refactor: composant AjouterJeuneButton
